requirements.txt
yt-dlp
ffmpeg
aria2p
pycryptodome